+++ R/peaks.R
@@ -199,7 +199,6 @@ CallPeaks.ChromatinAssay <- function(
   frags <- Fragments(object = object)
   # get all fragment file paths
   allfragpaths <- sapply(X = frags, FUN = GetFragmentData, slot = "path")
-  allfragpaths <- Reduce(f = paste, x = allfragpaths)
   gr <- CallPeaks(
     object = allfragpaths,
     macs2.path = macs2.path,
@@ -354,7 +353,7 @@ CallPeaks.default <- function(
     )
   }
 
-  gr <- makeGRangesFromDataFrame(df = df, keep.extra.columns = TRUE)
+  gr <- makeGRangesFromDataFrame(df = df, keep.extra.columns = TRUE, starts.in.df.are.0based = TRUE)
   if (cleanup) {
     files.to.remove <- paste0(outdir, .Platform$file.sep, files.to.remove)
     for (i in files.to.remove) {