+++ NEWS.md
@@ -6,6 +6,7 @@ Bug fixes:
 * Add check for correct number of columns in fragment file ([#748](https://github.com/timoast/signac/issues/748))
 * Fixed gene lookup when annotations contain NA values ([#771](https://github.com/timoast/signac/issues/771))
 * Fixed error in `ClosestFeature()` when query contained regions on contigs not present in gene annotation ([#758](https://github.com/timoast/signac/issues/758))
+* Fixed bug in `TSSEnrichment()` when using multiple fragment files ([#783](https://github.com/timoast/signac/issues/783))
 
 Other changes: