+++ NEWS.md
@@ -7,6 +7,7 @@ Bug fixes:
 * Fixed gene lookup when annotations contain NA values ([#771](https://github.com/timoast/signac/issues/771))
 * Fixed error in `ClosestFeature()` when query contained regions on contigs not present in gene annotation ([#758](https://github.com/timoast/signac/issues/758))
 * Fixed bug in `TSSEnrichment()` when using multiple fragment files ([#783](https://github.com/timoast/signac/issues/783))
+* Fixed bug in `CallPeaks()` when multiple fragment files used as input
 
 Other changes: